% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihoods_and_gradients.R
\name{as.gam}
\alias{as.gam}
\title{Coversion of an efp fit to a gam fit for plotting}
\usage{
as.gam(object)
}
\arguments{
\item{object}{an efp fitted model}
}
\value{
gam type object
}
\description{
Coversion of an efp fit to a gam fit for plotting
}

