% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overdispersion_function_annotated_complete.R
\name{BICadj}
\alias{BICadj}
\title{Adjusted BIC function}
\usage{
BICadj(model, data, overdispersion.output)
}
\arguments{
\item{model}{a fitted ef model}

\item{data}{the data used to fit the model}

\item{overdispersion.output}{output form the function \link{overdispersion}}
}
\value{
the adjusted BIC
}
\description{
Complete function for returning overdispersion estimates
}
\note{
When calling the function, you need to specify the data source for
  the model so that the number of site visits can be determined.
  You also need to specify the output from the overdispersion model
  to get the measure
}

