% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef-package.R
\docType{package}
\name{ef-package}
\alias{ef-package}
\alias{ef}
\title{Modelling Framework for the Estimation of Salmonid Abundance in Scottish Rivers}
\description{
A set of functions to estimate capture probabilities and densities
from multipass pass removal data.
}
