% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{ef}
\alias{ef}
\alias{ef-package}
\title{Modelling Framework for the Estimation of Salmonid Abundance in Scottish Rivers}
\description{
Modelling Framework for the Estimation of Salmonid Abundance in Scottish Rivers
}

