% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ef_data}
\alias{ef_data}
\title{Counts of salmon fry and parr from electrofishing.}
\format{
A data frame with 90 rows and 9 variables:
\describe{
\item{siteID}{description ...}
\item{year}{description ...}
\item{date}{description ...}
\item{EFpasscount:}{description ...}
\item{pass}{description ...}
\item{species}{description ...}
\item{lifestage}{description ...}
\item{count}{description ...}
\item{area}{description ...}
}
}
\source{
\url{https://www2.gov.scot/Topics/marine/Salmon-Trout-Coarse/Freshwater/Monitoring/temperature}
}
\usage{
ef_data
}
\description{
A dataset containing counts from multipass electrofishing
samples over three sites and 5 years.
}
\keyword{datasets}
