% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Likelihoods_and_gradients.R
\name{getData}
\alias{getData}
\title{Calculate the relavent statistics for fitting electrifishing models}
\usage{
getData(data, passnames = paste0("S0_R", 1:6))
}
\arguments{
\item{data}{a data.frame containing all relavent info.}

\item{passnames}{a character vector giving the names of the columns in which
the catch data reside.  These must be ordered so that the
first comes first, etc.}
}
\value{
a data frame
}
\description{
Basically, only three bits of info are needed from every electrofishing
event: The total catch, T, and a kind of cumulative catch measure termed X
in Seber (1978)
}
\details{
The data.frame must have the number of passes stored in a colum called 'Runs'
And fishing area must be in a row called 'Area'
}

