% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{invlogit}
\alias{invlogit}
\title{Inverse logistic transformation}
\usage{
invlogit(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
vector of values between 0 and 1
}
\description{
This function transforms values on the logistic scale to values on
the probability scale.
}

