% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{logit}
\alias{logit}
\title{Logistic transformation}
\usage{
logit(p)
}
\arguments{
\item{p}{a numeric vector with values between 0 and 1}
}
\value{
vector of values between -Inf and Inf
}
\description{
This function transforms values on the probability scale to values on
the logistic scale.
}
