% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transpar}
\alias{transpar}
\title{Utility function to convert parameters to probabilities}
\usage{
transpar(par, G)
}
\arguments{
\item{par}{fitted model parameters}

\item{G}{The design matrix for a model}
}
\value{
a data frame
}
\description{
The matrix G should be of dimension n x p,
and the parameter vector should be length p
}
