% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BICadj.R
\name{BICadj}
\alias{BICadj}
\title{Adjusted BIC function}
\usage{
BICadj(model, overdispersion.output)
}
\arguments{
\item{model}{a fitted ef model}

\item{overdispersion.output}{output from the function \link{overdispersion}}
}
\value{
the adjusted BIC
}
\description{
Complete function for returning overdispersion estimates
}
\note{
VisitID calculated from the data contained in the model object
}
