% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion.R
\name{overdispersion}
\alias{overdispersion}
\title{Estimating overdispersion}
\usage{
overdispersion(
  data,
  siteID = "siteID",
  visitID = "visitID",
  count = "count",
  pass = "pass",
  id = "sampleID",
  largemodel,
  control
)
}
\arguments{
\item{data}{dataframe containing EF data}

\item{siteID}{site name or unique ID}

\item{visitID}{a number identifying each unique site visit}

\item{count}{the number of fish caught for a particular combination of site visit,
species, lifestage and pass (defaults to "count")}

\item{pass}{EF pass number e.g. 1,2,3,4 (defaults to "pass")}

\item{largemodel}{passes control information to optimiser}

\item{sampleID}{sample ID i.e. unique combinations of site visit, species & lifestage}
}
\value{
a data.frame summarising overdispersion
}
\description{
Complete function for returning overdispersion estimates
}
\note{
ensure column names in function call are in inverted commas
}
