% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion.R
\name{overdispersion}
\alias{overdispersion}
\title{Estimating overdispersion}
\usage{
overdispersion(
  data,
  visitID = "visitID",
  count = "count",
  pass = "pass",
  sampleID = "sampleID",
  largemodel,
  control = "control"
)
}
\arguments{
\item{data}{dataframe containing EF data}

\item{visitID}{a number identifying each unique site visit}

\item{count}{the number of fish caught for a particular combination of site visit,
species, lifestage and pass (defaults to "count")}

\item{pass}{EF pass number e.g. 1,2,3,4 (defaults to "pass")}

\item{sampleID}{sample ID i.e. unique combinations of site visit, species & lifestage}

\item{largemodel}{a large model that captures most of the systematic
variation in the data - this is specified before
running the overdispersion function}

\item{control}{passes control information to optimiser}
}
\value{
a data.frame summarising overdispersion
}
\description{
Complete function for returning overdispersion estimates
}
\note{
ensure column names in function call are in inverted commas
}
